Add a casting specialization to hide G3SuperTimestream entirely from user space
Rearrange, no need for header
