+++ core/src/G3Frame.cxx
@@ -1,9 +1,6 @@
 #include <G3Frame.h>
 #include <G3Data.h>
 #include <G3Quat.h>
-#ifdef HAS_SUPERTIMESTREAM
-#include "G3SuperTimestream.h"
-#endif
 #include <serialization.h>
 #include <pybindings.h>
 
@@ -361,14 +358,6 @@ void G3Frame::blob_decode(struct blob_container &blob)
 	G3BufferInputStream is(*blob.blob);
 	cereal::PortableBinaryInputArchive item_ar(is);
 	item_ar >> make_nvp("val", ptr);
-
-#ifdef HAS_SUPERTIMESTREAM
-	// Convert to public type
-	auto v = std::dynamic_pointer_cast<G3SuperTimestream>(ptr);
-	if (!!v)
-		ptr = std::make_shared<G3TimestreamMap>(*v);
-#endif
-
 	blob.frameobject = ptr;
 
 	// Drop really big (> 128 MB) blobs at decode time. The savings
